Revert "Refactor the skills handler to remove bus logic for API  (#25446)"
This reverts commit 46258d7511784cd71617fdf0291571349f1a38ea.
