+++ front/pages/api/w/[wId]/skills/[sId]/index.ts
@@ -1,6 +1,5 @@
 /** @ignoreswagger */
 import { withSessionAuthenticationForWorkspace } from "@app/lib/api/auth_wrappers";
-import { AttachedKnowledgeSchema } from "@app/lib/api/skills/create_skill";
 import { resolveAdditionalRequestedSpaceModelIds } from "@app/lib/api/skills/space_requirements";
 import { type Authenticator, getFeatureFlags } from "@app/lib/auth";
 import { pruneOutdatedSkillEditSuggestions } from "@app/lib/reinforcement/skill_suggestion_pruning";
@@ -11,6 +10,7 @@ import { SkillResource } from "@app/lib/resources/skill/skill_resource";
 import { isResourceSId } from "@app/lib/resources/string_ids";
 import logger from "@app/logger/logger";
 import { apiError } from "@app/logger/withlogging";
+import { AttachedKnowledgeSchema } from "@app/pages/api/w/[wId]/skills";
 import type {
   SkillType,
   SkillWithRelationsType,
